+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes in 1.0-beta4
 
 * `@include` annotation replaces `@extends`.
-* Removed `@expectNothing`
+* Removed `@expectNothing`, `@expectException`, `@expectReturn` and `@expectReturnType` annotations
 * Now failes if examples produces outcomes that are not handled
 * `--filter` now only works on named examples
 * Added the `--flagged-only` option